% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithmetic.r
\docType{methods}
\name{arithmetic}
\alias{arithmetic}
\alias{+,shaq,numeric-method}
\alias{+,numeric,shaq-method}
\alias{-,shaq,numeric-method}
\alias{-,numeric,shaq-method}
\alias{*,shaq,numeric-method}
\alias{*,numeric,shaq-method}
\alias{/,shaq,numeric-method}
\alias{/,numeric,shaq-method}
\title{Arithmetic Operators}
\usage{
\S4method{+}{shaq,numeric}(e1, e2)

\S4method{+}{numeric,shaq}(e1, e2)

\S4method{-}{shaq,numeric}(e1, e2)

\S4method{-}{numeric,shaq}(e1, e2)

\S4method{*}{shaq,numeric}(e1, e2)

\S4method{*}{numeric,shaq}(e1, e2)

\S4method{/}{shaq,numeric}(e1, e2)

\S4method{/}{numeric,shaq}(e1, e2)
}
\arguments{
\item{e1, e2}{A shaq or a numeric vector.}
}
\value{
A shaq.
}
\description{
Some binary arithmetic operations for shaqs.  All operations are
vector-shaq or shaq-vector, but not shaq-shaq.  See details section for more
information.
}
\details{
Because the rules imposed on shaq objects are so minimal, it is next to
impossible to do some fairly basic things, like adding two shaqs together.
If you have two shaqs which are distributed \emph{identically} that you
wish to add, you can do so manually with something like
\code{DATA(c) <- Data(a) + Data(b)}
}
