% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bracket.r
\docType{methods}
\name{[,shaq-method}
\alias{[,shaq-method}
\title{subsetting}
\usage{
\S4method{[}{shaq}(x, i, j)
}
\arguments{
\item{x}{A shaq.}

\item{i, j}{Indices.  NOTE currently only implemented for \code{j} values.}
}
\value{
A shaq.
}
\description{
Subsetting via \code{`[`} for shaq objects.
}
