% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbind.r
\name{cbind.shaq}
\alias{cbind.shaq}
\title{cbind}
\usage{
cbind.shaq(..., deparse.level = 1)
}
\arguments{
\item{...}{A collection of shaqs.}

\item{deparse.level}{Ignored.}
}
\value{
A shaq.
}
\description{
Column binding for shaqs.
}
\details{
All shaqs should have the same number of rows.  Additionally, all shaqs
should be distributed in identical fashion.
}
\section{Communication}{

The operation is completely local.
}
\examples{
\dontrun{
library(kazaam)
x = ranshaq(runif, 10, 3)
y = ranshaq(runif, 10, 1)

cbind(x, y)

finalize()
}

}

