% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov.r
\name{cov}
\alias{cov}
\title{Covariance and Correlation}
\arguments{
\item{x}{A shaq.}

\item{y}{At this time, this must be \code{NULL}.}

\item{use}{NA handling rules, as with R's cov/cor functions.  At this time, only
"everything" is supported.}

\item{method}{The cov/cor method. Currently only "pearson" is available.}
}
\value{
A regular matrix.
}
\description{
Covariance and (pearson) correlation.
}
