% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossprod.r
\name{crossprod}
\alias{crossprod}
\alias{crossprod,shaq-method}
\alias{tcrossprod,tshaq-method}
\title{Matrix Multiplication}
\usage{
\S4method{crossprod}{shaq}(x, y = NULL)

\S4method{tcrossprod}{tshaq}(x, y = NULL)
}
\arguments{
\item{x}{A shaq.}

\item{y}{A shaq or \code{NULL}.}
}
\value{
A regular matrix.
}
\description{
Conceptually, this computes \code{t(x) \%*\% x} for a shaq \code{x}.
}
\section{Communication}{

The operation consists of a local crossproduct, followed by an
\code{allreduce()} call, quadratic on the number of columns.

For \code{crossprod()}, if the matrix distribution is poorly balanced
(specifically, if any rank has fewer rows than columns), then an inefficient
method is used. Similarly for \code{tcrossprod()} if the number of local rows
is greater than the number of local columns.
}

\examples{
\dontrun{
library(kazaam)
x = ranshaq(runif, 10, 3)

cp = crossprod(x)
comm.print(cp)

finalize()
}

}
