% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_local_dim.r
\name{get_local_dim}
\alias{get_local_dim}
\title{get_local_dim}
\usage{
get_local_dim(dim)
}
\arguments{
\item{dim}{Either the global number of rows or the global number of columns.}
}
\value{
The number of rows/columns which is local to the calling rank.
}
\description{
Given a global number of rows or columns, get the corresponding local number
of rows or columns. Pass the number of rows for a \code{shaq} and the number
of columns for a \code{tshaq}.
}
\details{
This assumes that the data is distributed across ranks in roughly equal
pieces. Modify the communicator by setting \code{.pbd_env$SPMD.CT$comm}.
}
