% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-getters.r
\docType{methods}
\name{getters}
\alias{getters}
\alias{nrow,shaq-method}
\alias{NROW,shaq-method}
\alias{ncol,shaq-method}
\alias{NCOL,shaq-method}
\alias{length,shaq-method}
\alias{Data}
\alias{Data,shaq-method}
\title{getters}
\usage{
\S4method{nrow}{shaq}(x)

\S4method{NROW}{shaq}(x)

\S4method{ncol}{shaq}(x)

\S4method{NCOL}{shaq}(x)

\S4method{length}{shaq}(x)

Data(x)

\S4method{Data}{shaq}(x)
}
\arguments{
\item{x}{A shaq.}
}
\description{
Getters for shaq objects.
}
