% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm.fit.r
\name{lm_coefs}
\alias{lm_coefs}
\title{lm_coefs}
\usage{
lm_coefs(x, y, tol = 1e-07)
}
\arguments{
\item{x}{A shaq (the input data).}

\item{y}{A shaq (the response).}
}
\value{
A regular vector.
}
\description{
Coefficients of the linear model.
}
