% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.r
\name{math}
\alias{math}
\alias{sqrt,shaq-method}
\alias{abs,shaq-method}
\alias{exp,shaq-method}
\alias{log,shaq-method}
\alias{log2,shaq-method}
\alias{log10,shaq-method}
\alias{sin,shaq-method}
\alias{cos,shaq-method}
\alias{tan,shaq-method}
\alias{asin,shaq-method}
\alias{acos,shaq-method}
\alias{atan,shaq-method}
\alias{sinh,shaq-method}
\alias{cosh,shaq-method}
\alias{tanh,shaq-method}
\alias{asinh,shaq-method}
\alias{acosh,shaq-method}
\alias{atanh,shaq-method}
\alias{ceiling,shaq-method}
\alias{floor,shaq-method}
\alias{trunc,shaq-method}
\alias{round,shaq-method}
\title{math}
\usage{
\S4method{sqrt}{shaq}(x)

\S4method{abs}{shaq}(x)

\S4method{exp}{shaq}(x)

\S4method{log}{shaq}(x, base = exp(1))

\S4method{log2}{shaq}(x)

\S4method{log10}{shaq}(x)

\S4method{sin}{shaq}(x)

\S4method{cos}{shaq}(x)

\S4method{tan}{shaq}(x)

\S4method{asin}{shaq}(x)

\S4method{acos}{shaq}(x)

\S4method{atan}{shaq}(x)

\S4method{sinh}{shaq}(x)

\S4method{cosh}{shaq}(x)

\S4method{tanh}{shaq}(x)

\S4method{asinh}{shaq}(x)

\S4method{acosh}{shaq}(x)

\S4method{atanh}{shaq}(x)

\S4method{ceiling}{shaq}(x)

\S4method{floor}{shaq}(x)

\S4method{trunc}{shaq}(x, ...)

\S4method{round}{shaq}(x, digits = 0)
}
\arguments{
\item{x}{A shaq or a numeric vector.}

\item{base}{Logarithm base.}

\item{...}{Additional arguments.}

\item{digits}{Number of decimal places to use.}
}
\value{
A shaq.
}
\description{
Basic math operations, like trig, rounding, exp/logs.
}
\details{
In each case, the corresponding base R function is called on the local data.
}
\section{Communication}{

Each operation is completely local.
}

\examples{
\dontrun{
library(kazaam)
x = ranshaq(runif, 10, 3)
y = ranshaq(runif, 10, 3)

sqrt(x)
log(y)

finalize()
}

}
