% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matmult.r
\docType{methods}
\name{matmult}
\alias{matmult}
\alias{\%*\%,shaq,matrix-method}
\title{Matrix Multiplication}
\usage{
\S4method{\%*\%}{shaq,matrix}(x, y)
}
\arguments{
\item{x}{A shaq.}

\item{y}{A regular matrix.}
}
\value{
A shaq.
}
\description{
Multiplies two distributed matrices, if they are conformable.
}
\details{
TODO
}
\examples{
\dontrun{
TODO
}

}
