% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.r
\docType{methods}
\name{norm}
\alias{norm}
\alias{norm,shaq,ANY-method}
\title{norm}
\usage{
\S4method{norm}{shaq,ANY}(x, type = c("O", "I", "F", "M", "2"))
}
\arguments{
\item{x}{A shaq}

\item{type}{The type of norm: one, infinity, frobenius, max-modulus, and spectral.}
}
\value{
A number (length 1 regular vector).
}
\description{
Implementation of R's \code{norm()} function for shaq objects.
}
\details{
TODO
}
\section{Communication}{

TODO
}

\examples{
\dontrun{
TODO
}

}
