% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbdSHAQ-package.r
\docType{package}
\name{pbdSHAQ-package}
\alias{pbdSHAQ-package}
\title{Tall Matrices}
\description{
TODO
}
\details{
This package requires an MPI library (OpenMPI, MPICH2, or LAM/MPI).
}
\references{
Programming with Big Data in R Website: \url{http://r-pbd.org/}
}
\author{
Drew Schmidt \email{wrathematics@gmail.com}, Wei-Chen Chen, Mike Matheson, and George Ostrouchov.
}
\keyword{Package}
