% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.r
\docType{methods}
\name{prcomp}
\alias{prcomp}
\alias{prcomp}
\alias{prcomp,shaq-method}
\title{Principal Components Analysis}
\usage{
prcomp(x, ...)

\S4method{prcomp}{shaq}(x, retx = TRUE, center = TRUE, scale. = FALSE,
  tol = NULL, ...)
}
\arguments{
\item{x}{A shaq.}

\item{...}{Ignored.}

\item{retx}{Should the rotated variables be returned?}

\item{center}{Should columns are zero centered?}

\item{scale.}{Should columns are rescaled to unit variance?}

\item{tol}{Ignored.}
}
\value{
A list of elements \code{sdev}, \code{rotation}, \code{center}, \code{scale},
and \code{x}, as with R's own \code{prcomp()}.  The elements are,
respectively, a regular vector, a regular matrix, a regular vector, a regular
vector, and a shaq.
}
\description{
Performs the principal components analysis.
}
\details{
\code{prcomp()} performs the principal components analysis on the data
matrix by taking the SVD. Sometimes core R and pbdSHAQ will disagree
slightly in what the rotated variables are because of how the SVD is
caluclated.
}
