% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.r
\docType{methods}
\name{print}
\alias{print}
\alias{print,shaq-method}
\alias{show,shaq-method}
\title{print}
\usage{
\S4method{print}{shaq}(x, ...)

\S4method{show}{shaq}(object)
}
\arguments{
\item{x, object}{A shaq.}

\item{...}{Ignored}
}
\description{
Print method for a shaq.
}
