% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qr.r
\name{qr}
\alias{qr}
\alias{qr_R}
\alias{qr_Q}
\title{QR Decomposition Methods}
\usage{
qr_R(x)

qr_Q(x, R)
}
\arguments{
\item{x}{A shaq.}

\item{R}{A regular matrix. This argument is optional, in that if it is not supplied
explicitly, then it will be computed in the background.  So if you already
have R lying around, supplying it here will improve performance.}
}
\value{
Q (a shaq) or R (a regular matrix).
}
\description{
QR Decomposition Methods
}
