% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_measure.r
\name{rand_measure}
\alias{rand_measure}
\title{rand_measure}
\usage{
rand_measure(l1, l2)
}
\arguments{
\item{l1, l2}{Two vectors or shaqs (with 1 column each) of the same length. If shaqs, they
must be distributed in the exact same way.}
}
\value{
The rand measure (global).
}
\description{
Returns the Rand measure, which is a statistic for comparing two data
clusterings.
}
\references{
Rand Index \url{https://en.wikipedia.org/wiki/Rand_index}
}
