% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranshaq.r
\name{ranshaq}
\alias{ranshaq}
\title{ranshaq}
\usage{
ranshaq(generator, nrows, ncols, local = FALSE, ...)
}
\arguments{
\item{generator}{}

\item{nrows, ncols}{}

\item{...}{Additional arguments passed to the generator.}
}
\value{
A shaq.
}
\description{
ranshaq
}
