% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcond.r
\docType{methods}
\name{rcond}
\alias{rcond}
\alias{rcond,shaq-method}
\title{rcond}
\usage{
\S4method{rcond}{shaq}(x, norm = c("O", "I", "1"), triangular = FALSE, ...)
}
\arguments{
\item{x}{A shaq.}

\item{norm, triangular, ...}{Arguments passed to \code{base::rcond()}.}
}
\value{
A number.
}
\description{
Reciprocal condition number estimate.
}
\details{
The estimate is computed by first forming the R matrix from a QR.  Currently
this involves first computing a crossproduct, so the estimate can be thought
of as being fairly liberal.  Afterwards, the condition number is estimated by
calling \code{base::rcond()} on the local R matrix.
}
\section{Communication}{

The operation is completely local except for forming the crossproduct, which
is an \code{allreduce()} call, quadratic on the number of columns.
}

\examples{
\dontrun{
library(kazaam)
x = ranshaq(runif, 10, 3)
rc = rcond(x)
comm.print(rc)

x = expand(matrix(1:30, 10))
rc = rcond(x)
comm.print(rc)

finalize()
}

}
