% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsvd.r
\name{rsvd}
\alias{rsvd}
\title{rsvd}
\usage{
rsvd(x, k = 1, q = 2, retu = TRUE, retv = TRUE)
}
\arguments{
\item{x}{The input data matrix.}

\item{k}{The number of singular values and/or left/right singular vectors
to estimate.}

\item{q}{An integer exponent, say 1, 2, or 3.  See the paper for details.}

\item{retu}{Logical; should the left singular vectors ("U") be returned?}

\item{retv}{Logical; should the right singular vectors ("V") be returned?}
}
\value{
A list of elements \code{d}, \code{u}, and \code{v}, as with R's own
\code{svd()}.  The elements are, respectively, a regular vector, a shaq, and
a regular matrix.
}
\description{
Randomized singular value decomposition.
}
\details{
A randomized SVD is an approximate method for computing using random
projections to estimate singular values/vectors. It is usually faster than
a full SVD (say via \code{svd()}) when only the first few singular
values/vectors are needed.
}
\section{Communication}{

The operation is completely local except for forming the crossproduct, which
is an \code{allreduce()} call, quadratic on the number of columns.
}

\examples{
\dontrun{
library(kazaam)
x = ranshaq(runif, 10, 3)

svd = rsvd(x)
comm.print(svd$d) # a globally owned vector
svd$u # a shaq
comm.print(svd$v) # a globally owned matrix

finalize()
}

}
\references{
Halko, Martinsson, and Tropp. 2011. Finding structure with randomness:
probabilistic algorithms for constructing approximate matrix decompositions.
SIAM Review 53 217-288.
}
