% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-setters.r
\docType{methods}
\name{setters}
\alias{setters}
\alias{Data<-}
\alias{Data<-,shaq-method}
\alias{DATA<-}
\alias{DATA<-,shaq-method}
\title{setters}
\usage{
Data(x, Data, checks = TRUE) <- value

\S4method{Data}{shaq}(x, Data, checks = TRUE) <- value

DATA(x, Data) <- value

\S4method{DATA}{shaq}(x, Data) <- value
}
\arguments{
\item{x}{A shaq.}

\item{Data}{The new data}
}
\description{
Setter functions for shaq objects.
}
\details{
\code{Data<-} will perform checks on the inserted data and ensure that the
number of columns match across processors (requiring communication).  It will
also udpate the number of rows as necessary.

\code{DATA<-} will perform no checks, so use with caution.
}
