% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-classes.r
\docType{class}
\name{shaq-class}
\alias{shaq-class}
\title{Class shaq}
\description{
Distributed tall matrix class.
}
\section{Slots}{

\describe{
\item{\code{DATA}}{The local submatrix.}

\item{\code{nrows,}}{ncols
The global matrix dimension.}
}}

\keyword{Classes}
