% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum.r
\name{sum}
\alias{sum}
\alias{sum,shaq-method}
\title{sum}
\usage{
\S4method{sum}{shaq}(x, ..., na.rm = FALSE)
}
\arguments{
\item{x}{A shaq.}

\item{...}{Additionaly elements (shaqs or numeric vectors/matrices) to sum.  Regular
vectors/matrices are assumed to be global, but will only be included to the
sum on rank 0.}

\item{na.rm}{should NA's be removed?}
}
\value{
A single value.
}
\description{
Sums the elements of a shaq.
}
\details{
Combinations of shaq and numeric matrices/vectors can be used, so long as the
first item is always a shaq.
}
\section{Communication}{

Communication consists of a an allreduce on a single value.
}

\examples{
\dontrun{
library(kazaam)
x = ranshaq(runif, 10, 3)

sum(x)
sum(x, 1)

finalize()
}

}
