% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svd.r
\docType{methods}
\name{svd,shaq-method}
\alias{svd,shaq-method}
\title{svd}
\usage{
\S4method{svd}{shaq}(x, nu = min(n, p), nv = min(n, p), LINPACK = FALSE)
}
\arguments{
\item{x}{A shaq.}

\item{nu}{number of left singular vectors to return.}

\item{nv}{number of right singular vectors to return.}

\item{LINPACK}{Ignored.}
}
\value{
TODO
}
\description{
Singular value decomposition.
}
