% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t.r
\name{t}
\alias{t}
\alias{t.shaq}
\alias{t.tshaq}
\title{t}
\usage{
t.shaq(x)

t.tshaq(x)
}
\arguments{
\item{x}{A shaq or tshaq.}
}
\value{
A tshaq or shaq (opposite of what was input).
}
\description{
Transpose a shaq or tshaq distributed matrix.
}
\section{Communication}{

The operation is completely local.
}
\examples{
\dontrun{
library(kazaam)
x = ranshaq(runif, 10, 3)

t(x)

finalize()
}

}

