% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bracket.r
\docType{methods}
\name{bracket}
\alias{bracket}
\alias{[,shaq-method}
\alias{[<-,shaq-method}
\title{subsetting}
\usage{
\S4method{[}{shaq}(x, i, j)

\S4method{[}{shaq}(x, i, j, ...) <- value
}
\arguments{
\item{x}{A shaq.}

\item{i, j}{Indices.  NOTE currently only implemented for \code{j} values.}

\item{...}{Ignored.}

\item{value}{Replacement value(s) for the \code{[<-} method.  This can either be an
appropriately sized numeric value or a shaq.  See the details section for
more information.}
}
\value{
A shaq.
}
\description{
Subsetting via \code{`[`} for shaq objects.
}
\details{
TODO
}
\section{Communication}{

TODO
}

\examples{
\dontrun{
TODO
}

}
