% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_ops.r
\name{col_ops}
\alias{col_ops}
\title{Column Operations}
\arguments{
\item{x}{A shaq.}

\item{na.rm}{Should \code{NA}'s be removed?}

\item{dims}{Ignored.}
}
\value{
A regular vector.
}
\description{
Column operations (currently sums/means) for shaq objects.
}
