% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.r
\name{collapse}
\alias{collapse}
\title{collapse}
\usage{
collapse(x)
}
\arguments{
\item{x}{A shaq.}
}
\value{
A regular matrix (rank 0) or \code{NULL} (everyone else).
}
\description{
Collapse a shaq into a regular matrix.
}
\details{
Only rank 0 will own the matrix on return.
}
\section{Communication}{

Short answer: quite a bit.  Each local submatrix has to be sent to rank 0.
}

\examples{
\dontrun{
library(kazaam)
dx = ranshaq(runif, 10, 3)

x = collapse(dx)
comm.print(x)

finalize()
}

}
