% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand.r
\name{expand}
\alias{expand}
\title{expand}
\usage{
expand(x)
}
\arguments{
\item{x}{A regular matrix.}
}
\value{
A shaq.
}
\description{
Expand a regular matrix owned on MPI rank 0 into a shaq.
}
\examples{
\dontrun{
TODO
}

}
