% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-getters.r
\docType{methods}
\name{getters}
\alias{Data}
\alias{Data,shaq-method}
\alias{NCOL,shaq-method}
\alias{NROW,shaq-method}
\alias{getters}
\alias{length,shaq-method}
\alias{ncol,shaq-method}
\alias{ncol.local}
\alias{ncol.local,shaq-method}
\alias{nrow,shaq-method}
\alias{nrow.local}
\alias{nrow.local,shaq-method}
\title{getters}
\usage{
\S4method{nrow}{shaq}(x)

\S4method{NROW}{shaq}(x)

nrow.local(x)

\S4method{nrow.local}{shaq}(x)

\S4method{ncol}{shaq}(x)

\S4method{NCOL}{shaq}(x)

ncol.local(x)

\S4method{ncol.local}{shaq}(x)

\S4method{length}{shaq}(x)

Data(x)

\S4method{Data}{shaq}(x)
}
\arguments{
\item{x}{A shaq.}
}
\description{
Getters for shaq objects.
}
\details{
Functions to return the number of rows (\code{nrow()} and \code{NROW()}),
the number of columns (\code{ncol()} and \code{NCOL()}), the length - or
product of the number of rows and cols - (\code{length()}), and the local
submatrix (\code{Data()}).
}
\section{Communication}{

Each operation is completely local.
}
\seealso{
\code{\link{setters}}
}

