% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.shaq.r
\name{is.shaq}
\alias{is.shaq}
\title{is.shaq}
\usage{
is.shaq(x)
}
\arguments{
\item{x}{An R object.}
}
\value{
A logical value, indicating whether or not the input is a shaq.
}
\description{
Test if an object is a shaq.
}
\section{Communication}{

The operation is completely local.
}
\examples{
\dontrun{
library(kazaam)
x = ranshaq(runif, 10, 3)

comm.print(is.shaq(x))

finalize()
}

}

