% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matmult.r
\docType{methods}
\name{matmult}
\alias{matmult}
\alias{\%*\%,shaq,matrix-method}
\title{Matrix Multiplication}
\usage{
\S4method{\%*\%}{shaq,matrix}(x, y)
}
\arguments{
\item{x}{A shaq.}

\item{y}{A regular matrix, globally owned on all ranks.  Since the number of columns
of a shaq should be small, this matrix should be small as well.}
}
\value{
A shaq.
}
\description{
Multiplies two distributed matrices, if they are conformable.
}
\details{
The two shaqs must be distributed \emph{identically}.
}
\section{Communication}{

The operation is completely local.
}

\examples{
\dontrun{
library(kazaam)
x = ranshaq(runif, 10, 3)
y = matrix(1:9, 3, 3)

x \%*\% y

finalize()
}

}
