% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.r
\docType{methods}
\name{prcomp}
\alias{prcomp}
\alias{prcomp}
\alias{prcomp,shaq-method}
\title{Principal Components Analysis}
\usage{
prcomp(x, ...)

\S4method{prcomp}{shaq}(x, retx = TRUE, center = TRUE, scale. = FALSE,
  tol = NULL, ...)
}
\arguments{
\item{x}{A shaq.}

\item{...}{Ignored.}

\item{retx}{logical value, indicates whether the rotated variables should be returned}

\item{center}{logical value, determines whether or not columns are zero centered}

\item{scale.}{logical value, determines whether or not columns are rescaled to unit variance}

\item{tol}{a value indicating the magnitude below which components should be
omitted. (Components are omitted if their standard deviations are less than
or equal to \code{tol} times the standard deviation of the first component.)
With the default null setting, no components are omitted.  Other settings
for tol could be \code{tol = 0} or \code{tol = sqrt(.Machine$double.eps)},
which would omit essentially constant components}
}
\value{
TODO
}
\description{
Performs the principal components analysis.
}
\details{
\code{prcomp()} performs the principal components analysis on the data
matrix by taking the SVD. Sometimes core R and pbdSHAQ will disagree
slightly in what the rotated variables are because of how the SVD is
caluclated.
}
