% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.r
\docType{methods}
\name{print}
\alias{print}
\alias{print,gbd1d-method}
\alias{show,gbd1d-method}
\title{print}
\usage{
\S4method{print}{gbd1d}(x, ...)

\S4method{show}{gbd1d}(object)
}
\arguments{
\item{x, object}{A shaq.}

\item{...}{Ignored}
}
\description{
Print method for a shaq.
}
\section{Communication}{

The operation is completely local.
}
\examples{
\dontrun{
library(kazaam)
x = shaq(1, 10, 3)

x # same as print(x) or comm.print(x)

finalize()
}

}

