% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.r
\docType{methods}
\name{scale}
\alias{scale}
\alias{scale}
\alias{scale,shaq,logical,logical-method}
\title{Scale}
\usage{
scale(x, center = TRUE, scale = TRUE)

\S4method{scale}{shaq,logical,logical}(x, center = TRUE, scale = TRUE)
}
\arguments{
\item{x}{A shaq.}

\item{center}{logical value, determines whether or not columns are zero centered}

\item{scale}{logical value, determines whether or not columns are rescaled to unit variance}
}
\value{
A shaq.
}
\description{
Centers and/or scales the columns of a distributed matrix.
}
\section{Communication}{

The communication consists of two allreduce calls, each quadratic on the
number of columns.
}

\examples{
\dontrun{
library(kazaam)
x = ranshaq(rnorm, 10, 3, mean=30, sd=10)

x
scale(x)

finalize()
}

}
