% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.r
\name{scale}
\alias{scale}
\title{Scale}
\arguments{
\item{x}{A shaq.}

\item{center}{logical value, determines whether or not columns are zero centered}

\item{scale}{logical value, determines whether or not columns are rescaled to unit variance}
}
\value{
Returns a distributed matrix.
}
\description{
Centers and/or scales the columns of a distributed matrix.
}
