% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-setters.r
\docType{methods}
\name{setters}
\alias{DATA<-}
\alias{DATA<-,shaq-method}
\alias{setters}
\title{setters}
\usage{
DATA(x) <- value

\S4method{DATA}{shaq}(x) <- value
}
\arguments{
\item{x}{A shaq.}

\item{value}{The new data.}
}
\description{
Setter functions for shaq objects.  Generally not recommended unless you are
sure you know what you're doing.
}
\details{
\code{Data<-} will perform checks on the inserted data and ensure that the
number of columns match across processors (requiring communication).  It will
also udpate the number of rows as necessary.

\code{DATA<-} will perform no checks, so use only if you're really sure that
you know what you're doing.
}
\section{Communication}{

With \code{Data<-}, a check on the global number of rows is performed.  This
amounts to an allgather operation on a logical value (the local dimension
check).
}
\seealso{
\code{\link{getters}}, \code{\link{bracket}}
}

