% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-constructor.r
\name{shaq}
\alias{shaq}
\alias{shaq.NULL}
\alias{shaq.matrix}
\alias{shaq.numeric}
\title{shaq}
\usage{
shaq(Data, nrows, ncols, checks = TRUE)

\method{shaq}{matrix}(Data, nrows, ncols, checks = TRUE)

\method{shaq}{numeric}(Data, nrows, ncols, checks = TRUE)

\method{shaq}{NULL}(Data, nrows, ncols, checks = TRUE)
}
\arguments{
\item{Data}{The local submatrix.}

\item{nrows, ncols}{The GLOBAL number of rows and columns.}

\item{checks}{Logical. Should some basic dimension checks be performed?  Note that these
require communication, and with many MPI ranks, could be expensive.}
}
\description{
Constructor for shaq objects.
}
\details{
If \code{nrows} and/or \code{ncols} is missing, then it will be imputed.
One can pass \code{NULL} for the \code{Data} argument to specify that that
MPI rank owns no data. In this case, you \emph{must} manually provide
\code{ncols}. This use case is typical when reading from a subset of
processors and then broadcasting out to the remainder.
}
\section{Communication}{

If \code{checks=TRUE}, a check on the global number of rows is performed.
This amounts to an allgather operation on a logical value (the local
dimension check).
}
\seealso{
\code{\link{shaq-class}} and \code{\link{load_balance}}
}

