% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-constructor.r
\name{shaq}
\alias{shaq}
\alias{tshaq}
\alias{shaq.matrix}
\alias{shaq.float32}
\alias{shaq.vector}
\alias{shaq.integer}
\alias{shaq.double}
\alias{shaq.NULL}
\alias{tshaq.matrix}
\alias{tshaq.float32}
\alias{tshaq.vector}
\alias{tshaq.NULL}
\title{shaq}
\usage{
shaq(Data, nrows, ncols, checks = TRUE)

tshaq(Data, nrows, ncols, checks = TRUE)

\method{shaq}{matrix}(Data, nrows, ncols, checks = TRUE)

\method{shaq}{float32}(Data, nrows, ncols, checks = TRUE)

\method{shaq}{vector}(Data, nrows, ncols, checks = TRUE)

\method{shaq}{integer}(Data, nrows, ncols, checks = TRUE)

\method{shaq}{double}(Data, nrows, ncols, checks = TRUE)

\method{shaq}{NULL}(Data, nrows, ncols, checks = TRUE)

\method{tshaq}{matrix}(Data, nrows, ncols, checks = TRUE)

\method{tshaq}{float32}(Data, nrows, ncols, checks = TRUE)

\method{tshaq}{vector}(Data, nrows, ncols, checks = TRUE)

\method{tshaq}{NULL}(Data, nrows, ncols, checks = TRUE)
}
\arguments{
\item{Data}{The local submatrix.}

\item{nrows, ncols}{The GLOBAL number of rows and columns.}

\item{checks}{Logical. Should some basic dimension checks be performed?  Note that these
require communication, and with many MPI ranks, could be expensive.}
}
\description{
Constructor for shaq objects.
}
\details{
If \code{nrows} and/or \code{ncols} is missing, then it will be imputed.
The internal logic tries closely to mimic R's, but it may do unexpected
things that you do not want. You are encouraged to specify the distributed
dimension (i.e., \code{nrows} for a \code{shaq}, \code{ncols} for
\code{tshaq}).

One can pass \code{NULL} for the \code{Data} argument to specify that that
MPI rank owns no data. In this case, you \emph{must} manually provide the
non-distributed dimension (i.e., \code{ncols} for \code{shaq}). This use case
is typical when reading from a subset of processors and then broadcasting out
to the remainder.
}
\section{Communication}{

If \code{checks=TRUE}, a check on the global number of rows is performed.
This amounts to an allgather operation on a logical value (the local
dimension check).
}

\seealso{
\code{\link{shaq-class}} and \code{\link{load_balance}}
}
