% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-constructor.r
\name{shaq}
\alias{shaq}
\alias{shaq}
\title{shaq}
\usage{
shaq(Data = matrix(nrow = 0, ncol = 0), nrows = 0, ncols = 0,
  checks = TRUE)
}
\arguments{
\item{Data}{The local submatrix.}

\item{nrows, ncols}{The GLOBAL number of rows and columns.}

\item{checks}{Logical. Should some basic dimension checks be performed?  Note that these
require communication, and with many MPI ranks, could be expensive.}
}
\description{
Constructor for shaq objects.
}
